Add redirect to success page
To find the success page, look in views/products/ for success.html.erb.
There's a route for it that points to products#success, so the url can
just be /success. That's where it redirects in the controller.

If it fails, there's no redirect. The page currently renders a json
error message, but that can be changed.